refactor(folios): refactor get child folio card
