+++ tests/functional/controllers/folios_controller.spec.ts
@@ -269,117 +269,4 @@ test.group('Folio controller', (group) => {
     assert.equal(folio.statistics.cardMarketTrending, 'up') // 20 > 10
     assert.equal(folio.statistics.tcgPlayerTrending, 'up') // 15 > 8
   })
-
-  test('childFolioCards - should return paginated cards from specific child folio', async ({
-    client,
-    assert,
-  }) => {
-    const userId = TEST_AUTH_USER_ID
-
-    await ArtistFactory.create()
-    await RarityFactory.create()
-    await LegalityFactory.create()
-    await SetFactory.merge({ id: 'base1' }).create()
-
-    const childFolio = await FolioFactory.merge({
-      userId,
-      name: 'My Custom Collection',
-      isRoot: false,
-    }).create()
-
-    const otherChildFolio = await FolioFactory.merge({
-      userId,
-      name: 'Other Collection',
-      isRoot: false,
-    }).create()
-
-    const cards = await CardFactory.createMany(3)
-
-    // Add cards to target child folio
-    await CardFolioFactory.merge([
-      { cardId: cards[0].id, folioId: childFolio.id, occurrence: 2 },
-      { cardId: cards[1].id, folioId: childFolio.id, occurrence: 1 },
-    ]).createMany(2)
-
-    // Add card to other folio (should not be returned)
-    await CardFolioFactory.merge({
-      cardId: cards[2].id,
-      folioId: otherChildFolio.id,
-      occurrence: 3,
-    }).create()
-
-    const response = await client
-      .get(`/api/v1/folios/${childFolio.id}/cards`)
-      .header('Authorization', 'Bearer fake-token-for-testing')
-      .qs({ page: 1, limit: 10 })
-
-    response.assertStatus(200)
-
-    const body = response.body()
-    assert.properties(body, ['meta', 'data'])
-    assert.properties(body.meta, ['total', 'currentPage', 'perPage'])
-    assert.equal(body.meta.total, 2)
-    assert.equal(body.meta.currentPage, 1)
-
-    assert.isArray(body.data)
-    assert.lengthOf(body.data, 2)
-
-    const firstCardFolio = body.data[0]
-    assert.properties(firstCardFolio, ['id', 'occurrence', 'card'])
-    assert.properties(firstCardFolio.card, ['id', 'imageSmall'])
-
-    const cardIds = body.data.map((cardFolio: any) => cardFolio.card.id)
-    assert.includeMembers(cardIds, [cards[0].id, cards[1].id])
-    assert.notInclude(cardIds, cards[2].id) // Should not include card from other folio
-
-    const cardFolioWithOccurrence2 = body.data.find((cf: any) => cf.card.id === cards[0].id)
-    const cardFolioWithOccurrence1 = body.data.find((cf: any) => cf.card.id === cards[1].id)
-
-    assert.equal(cardFolioWithOccurrence2.occurrence, 2)
-    assert.equal(cardFolioWithOccurrence1.occurrence, 1)
-  })
-
-  test('childFolioCards - should return 403 when folio is not owned by user', async ({
-    client,
-  }) => {
-    const otherUserId = 'other-user-id'
-
-    const otherUserFolio = await FolioFactory.merge({
-      userId: otherUserId,
-      name: 'Other User Collection',
-      isRoot: false,
-    }).create()
-
-    const response = await client
-      .get(`/api/v1/folios/${otherUserFolio.id}/cards`)
-      .header('Authorization', 'Bearer fake-token-for-testing')
-      .qs({ page: 1, limit: 10 })
-
-    response.assertStatus(403)
-    response.assertBodyContains({
-      code: 'E_FOLIO_NOT_OWNED',
-    })
-  })
-
-  test('childFolioCards - should return 422 when trying to access main folio cards', async ({
-    client,
-  }) => {
-    const userId = TEST_AUTH_USER_ID
-
-    const mainFolio = await FolioFactory.merge({
-      userId,
-      name: 'root',
-      isRoot: true,
-    }).create()
-
-    const response = await client
-      .get(`/api/v1/folios/${mainFolio.id}/cards`)
-      .header('Authorization', 'Bearer fake-token-for-testing')
-      .qs({ page: 1, limit: 10 })
-
-    response.assertStatus(422)
-    response.assertBodyContains({
-      code: 'E_NOT_CHILD_FOLIO',
-    })
-  })
 })
